SMS sending completed

+++ database/migrations/2021_09_21_083809_add_balance_to_users.php
@@ -15,7 +15,7 @@ class AddBalanceToUsers extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             //
-            $table->integer('balance');
+            $table->double('balance', 8, 2);
         });
     }
